# pesitwizard-docs/public/api/openapi-client.yaml
openapi: 3.0.3
info:
  title: PeSIT Client API
  description: API pour l'envoi et la réception de fichiers via le protocole PeSIT
  version: 1.0.0
  contact:
    email: contact@pesit.cloud

servers:
  - url: http://localhost:9081
    description: Serveur local

tags:
  - name: Servers
    description: Gestion des serveurs PeSIT cibles
  - name: Transfers
    description: Envoi et réception de fichiers

paths:
  /api/v1/servers:
    get:
      tags: [Servers]
      summary: Liste des serveurs configurés
      responses:
        '200':
          description: Liste des serveurs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Server'
    post:
      tags: [Servers]
      summary: Créer un serveur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerCreate'
      responses:
        '201':
          description: Serveur créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'

  /api/v1/servers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Servers]
      summary: Détail d'un serveur
      responses:
        '200':
          description: Détail du serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '404':
          description: Serveur non trouvé
    put:
      tags: [Servers]
      summary: Modifier un serveur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerUpdate'
      responses:
        '200':
          description: Serveur modifié
    delete:
      tags: [Servers]
      summary: Supprimer un serveur
      responses:
        '204':
          description: Serveur supprimé

  /api/v1/servers/{id}/test:
    post:
      tags: [Servers]
      summary: Tester la connexion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Résultat du test
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionTest'

  /api/transfers:
    get:
      tags: [Transfers]
      summary: Liste des transferts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TransferStatus'
        - name: direction
          in: query
          schema:
            $ref: '#/components/schemas/TransferDirection'
        - name: serverId
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Liste paginée des transferts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferPage'

  /api/transfers/send:
    post:
      tags: [Transfers]
      summary: Envoyer un fichier
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - serverId
                - remoteFilename
                - partnerId
              properties:
                file:
                  type: string
                  format: binary
                  description: Fichier à envoyer
                serverId:
                  type: integer
                  description: ID du serveur cible
                partnerId:
                  type: string
                  description: Identifiant partenaire (PI 3 DEMANDEUR)
                password:
                  type: string
                  description: Mot de passe d'authentification (PI 5)
                remoteFilename:
                  type: string
                  description: Nom du fichier distant (PI 12)
                virtualFile:
                  type: string
                  description: Identifiant du fichier virtuel (optionnel)
      responses:
        '200':
          description: Transfert effectué
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        '400':
          description: Requête invalide
        '500':
          description: Erreur de transfert

  /api/transfers/receive:
    post:
      tags: [Transfers]
      summary: Recevoir un fichier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiveRequest'
      responses:
        '200':
          description: Transfert effectué
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'

  /api/transfers/{id}:
    get:
      tags: [Transfers]
      summary: Détail d'un transfert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Détail du transfert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'

  /api/transfers/{id}/download:
    get:
      tags: [Transfers]
      summary: Télécharger un fichier reçu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Fichier binaire
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

components:
  schemas:
    Server:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        host:
          type: string
        port:
          type: integer
        serverId:
          type: string
          description: Identifiant du serveur PeSIT (PI 4)
        tlsEnabled:
          type: boolean
        status:
          type: string
          enum: [CONNECTED, DISCONNECTED, ERROR]

    ServerCreate:
      type: object
      required:
        - name
        - host
        - port
        - serverId
      properties:
        name:
          type: string
          description: Nom du serveur (pour affichage)
        host:
          type: string
          description: Adresse IP ou hostname du serveur
        port:
          type: integer
          default: 6502
          description: Port PeSIT (défaut 6502)
        serverId:
          type: string
          description: Identifiant du serveur PeSIT (PI 4)
        tlsEnabled:
          type: boolean
          default: false

    ServerUpdate:
      type: object
      properties:
        name:
          type: string
        host:
          type: string
        port:
          type: integer
        password:
          type: string
        tlsEnabled:
          type: boolean

    ConnectionTest:
      type: object
      properties:
        success:
          type: boolean
        latency:
          type: integer
          description: Latence en ms
        serverVersion:
          type: integer
        message:
          type: string

    Transfer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/TransferStatus'
        direction:
          $ref: '#/components/schemas/TransferDirection'
        filename:
          type: string
        localPath:
          type: string
        size:
          type: integer
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        duration:
          type: integer
          description: Durée en ms
        error:
          type: string

    TransferStatus:
      type: string
      enum: [PENDING, IN_PROGRESS, COMPLETED, FAILED, CANCELLED]

    TransferDirection:
      type: string
      enum: [SEND, RECEIVE]

    TransferPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Transfer'
        totalElements:
          type: integer
        totalPages:
          type: integer
        number:
          type: integer
        size:
          type: integer

    ReceiveRequest:
      type: object
      required:
        - serverId
        - remoteFilename
        - partnerId
      properties:
        serverId:
          type: integer
          description: ID du serveur cible
        partnerId:
          type: string
          description: Identifiant partenaire (PI 3 DEMANDEUR)
        password:
          type: string
          description: Mot de passe d'authentification (PI 5)
        remoteFilename:
          type: string
          description: Nom du fichier distant (PI 12)
        virtualFile:
          type: string
          description: Identifiant du fichier virtuel (optionnel)
        filename:
          type: string
          description: Chemin local de destination